Projeto utilizado no curso Mocks em PHP: entenda os dublês de testes
