--- mocks-aula-5-completa/src/Service/Encerrador.php
@@ -1,34 +1,0 @@
-<?php
-
-namespace Alura\Leilao\Service;
-
-use Alura\Leilao\Dao\Leilao as LeilaoDao;
-
-class Encerrador
-{
-    private $dao;
-    private $enviadorEmail;
-
-    public function __construct(LeilaoDao $dao, EnviadorEmail $enviadorEmail)
-    {
-        $this->dao = $dao;
-        $this->enviadorEmail = $enviadorEmail;
-    }
-
-    public function encerra()
-    {
-        $leiloes = $this->dao->recuperarNaoFinalizados();
-
-        foreach ($leiloes as $leilao) {
-            if ($leilao->temMaisDeUmaSemana()) {
-                try {
-                    $leilao->finaliza();
-                    $this->dao->atualiza($leilao);
-                    $this->enviadorEmail->notificarTerminoLeilao($leilao);
-                } catch (\DomainException $e) {
-                    error_log($e->getMessage());
-                }
-            }
-        }
-    }
-}
