--- mocks-aula-5-completa/src/Service/EnviadorEmail.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace Alura\Leilao\Service;
-
-use Alura\Leilao\Model\Leilao;
-
-class EnviadorEmail
-{
-    public function notificarTerminoLeilao(Leilao $leilao): void
-    {
-        $sucesso = mail(
-            'usuario@email.com',
-            'Leilão finalizado',
-            'O leilão para ' . $leilao->recuperarDescricao() . ' foi finalizado'
-        );
-
-        if (!$sucesso) {
-            throw new \DomainException('Erro ao enviar e-mail');
-        }
-    }
-}
