+++ php/mocks-em-php/README.md
@@ -1,0 +1,13 @@
+# Mocks em PHP: entenda os dublês de testes
+
+Categoria | Carga horária | Atividades | Finalizado | Certificado |
+:-:|:-:|:-:|:-:|:-:|
+Programação | 10 horas | 40 | 27/02/2022 | [Certificado Normal](https://cursos.alura.com.br/certificate/84c51076-a163-4cb3-be7d-0461126be27c) / [Certificado Formal](https://cursos.alura.com.br/user/rodineicosta/course/php-mocks/formalCertificate)
+
+## Aulas
+
+- Isolando o DAO
+- Conhecendo Mocks e Stubs
+- Dublês atrapalhados
+- Verificações extras
+- Inimigos do teste
